Editor aufgeräumt: AdbEditor von 1000 auf 150 Zeilen gestutzt
- Metadaten (Typ/Autor/Lizenz) in EditorMetadataBar ausgegliedert
- DnD-Liste und XML-Editor in TemplateEditorPanel ausgegliedert
- Vorschau-Dialog in PreviewDialog ausgegliedert
